requirements.txt
Django~=4.0.3
django-timestampable[drf]~=1.0.3
django-taggit~=2.1.0
